requirements.txt
pygame
edge-tts
python-dotenv
playsound
AppOpener
pywhatkit
beautifulsoup4
rich
groq
requests
keyboard
cohere
googlesearch-python
